Rename: Move to constant package
From: com.fastcampus.investment.util.type
To : com.fastcampus.investment.constant
